Remove the developer preview opt-in
